+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0 ##
 * Split out schema validator to reduce bundle size
+* Handle readonly properties on array items
 
 ## 1.2 ##
 * Add `compile` and `loadCompiled` specs for faster validation